+++ sdmx-dl-format-xml/src/main/java/internal/sdmxdl/format/xml/XMLStreamFlow20.java
@@ -122,7 +122,7 @@ public final class XMLStreamFlow20 {
                 .builder()
                 .ref(flowRef)
                 .structureRef(structRef)
-                .label(flowName.build(id))
+                .name(flowName.build(id))
                 .description(flowDescription.build(""))
                 .build();
     }