+++ sdmx-dl-format-xml/src/test/java/internal/sdmxdl/format/xml/XMLStreamFlow21Test.java
@@ -45,19 +45,19 @@ public class XMLStreamFlow21Test {
                                 .builder()
                                 .ref(DataflowRef.of("ECB", "AME", "1.0"))
                                 .structureRef(DataStructureRef.of("ECB", "ECB_AME1", "1.0"))
-                                .label("AMECO")
+                                .name("AMECO")
                                 .build(),
                         Dataflow
                                 .builder()
                                 .ref(DataflowRef.of("ECB", "BKN", "1.0"))
                                 .structureRef(DataStructureRef.of("ECB", "ECB_BKN1", "1.0"))
-                                .label("Banknotes statistics")
+                                .name("Banknotes statistics")
                                 .build(),
                         Dataflow
                                 .builder()
                                 .ref(DataflowRef.of("ECB", "BLS", "1.0"))
                                 .structureRef(DataStructureRef.of("ECB", "ECB_BLS1", "1.0"))
-                                .label("Bank Lending Survey Statistics")
+                                .name("Bank Lending Survey Statistics")
                                 .build()
                 );
     }
@@ -72,14 +72,14 @@ public class XMLStreamFlow21Test {
                                 .builder()
                                 .ref(DataflowRef.of("CD2030", "CD2030", "1.0"))
                                 .structureRef(DataStructureRef.of("CD2030", "CD2030", "1.0"))
-                                .label("Coundown 2030")
+                                .name("Coundown 2030")
                                 .description("This dataset is used to support the downolad of the CD2030 data")
                                 .build(),
                         Dataflow
                                 .builder()
                                 .ref(DataflowRef.of("EAPRO", "DF_EAPRO_CROSS_SECTOR", "1.0"))
                                 .structureRef(DataStructureRef.of("EAPRO", "DSD_EAPRO", "1.0"))
-                                .label("EAPRO Cross Sector Indicators")
+                                .name("EAPRO Cross Sector Indicators")
                                 .description("A dataflow based on a subset of the global cross-sector indicators.")
                                 .build()
                 );