+++ sdmx-dl-provider-ri/src/main/java/internal/sdmxdl/provider/ri/web/drivers/PxWebDriver.java
@@ -448,7 +448,7 @@ public final class PxWebDriver implements WebDriver {
                     .builder()
                     .ref(DataflowRef.of(dbId, id, null))
                     .structureRef(DataStructureRef.of(dbId, id, null))
-                    .label(title)
+                    .name(title)
                     .build();
         }