+++ sdmx-dl-provider-ri/src/main/java/internal/sdmxdl/provider/ri/web/drivers/RngDriver.java
@@ -113,7 +113,7 @@ public final class RngDriver implements WebDriver {
 
         @Override
         public @NonNull Collection<Dataflow> getFlows() {
-            return singleton(Dataflow.builder().ref(DataflowRef.parse("RNG")).structureRef(DataStructureRef.parse("STRUCT_RNG")).label("RNG").build());
+            return singleton(Dataflow.builder().ref(DataflowRef.parse("RNG")).structureRef(DataStructureRef.parse("STRUCT_RNG")).name("RNG").build());
         }
 
         @Override