+++ CHANGELOG.md
@@ -9,12 +9,14 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- ![API] Add description field in `Dataflow` [#287](https://github.com/nbbrd/sdmx-dl/issues/287)
+- ![API] Add field `Dataflow#description` [#287](https://github.com/nbbrd/sdmx-dl/issues/287)
 - ![SOURCE] Add source from UN International Children’s Emergency Fund [#95](https://github.com/nbbrd/sdmx-dl/issues/95)
 
 ### Changed
 
 - ![API] Improve code coherence by replacing `Dataflow#of(...)` and `DataQuery#of(...)` with builders
+- ![API] Rename field `Dataflow#label` as `Dataflow#name`
+- ![CLI] Modify `list/flows` command headers to follow `Dataflow` changes
 - ![DOC] Enforce https in doc URL
 
 ### Fixed